Incremental reimplementation of RasterSource reading to depend on geotrellis-contrib version. Linting fixes/cleanup.

+++ core/src/main/scala/astraea/spark/rasterframes/expressions/TileAssembler.scala
@@ -118,7 +118,7 @@ case class TileAssembler(
     val cells = Array.ofDim[Double](length)
     result.get(cells)
     val (tileCols, tileRows) = buffer.tileSize
-    val tile = ArrayTile(cells, tileCols, tileRows)
+    val tile = ArrayTile(cells, tileCols.toInt, tileRows.toInt)
     TileType.serialize(tile)
   }
